pyproject.toml
[project]
name = "booking"
version = "0.1.0"
description = "Scrape data from Booking.com"
readme = "README"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "llama-index-graph-stores-neo4j>=0.4.4",
    "llama-index>=0.12.8",
    "pyarrow>=18.1.0",
    "pydantic>=2.10.4",
    "python-dateutil>=2.9.0.post0",
    "requests>=2.32.3",
    "typing-extensions>=4.12.2",
    "numpy==1.24.4",
    "scipy==1.12.0",
    "graspologic>=3.3.0",
    "future>=1.0.0",
    "pyvis>=0.3.2",
    "python-louvain>=0.16",
    "llama-index-llms-groq>=0.3.1",
    "torch>=2.5.1",
    "neo4j>=5.27.0",
    "scikit-learn>=1.6.0",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "ruff>=0.8.4",
]

[tool.ruff]
exclude = ["*.ipynb"]

[tool.uv]
package = true

[build-system]
requires = ["setuptools", "wheel", "pip"]
build-backend = "setuptools.build_meta"
